fix(jit): cast std::pow result to float to suppress C4244 warning
【动机】
计算 spillWeight 时使用 `std::pow(float, int)`，返回值类型为 double，赋值给 float 变量时触发 MSVC C4244 警告（在 /WX 模式下被视为错误）。需显式转换以消除编译警告。

【原理】
将 `std::pow` 的结果用 `static_cast<float>` 显式转换为 float，消除精度丢失警告，不影响数值逻辑。

+++ src/jit/pass/LivenessAnalysis.h
@@ -174,7 +174,7 @@ private:
                     if (def.isVirtual()) {
                         live.erase(def.vreg());
                         LiveInterval& interval = getInterval(def.vreg());
-                        interval.spillWeight += std::pow(10.0f, block->loopDepth());
+                        interval.spillWeight += static_cast<float>(std::pow(10.0, block->loopDepth()));
                         if (!interval.ranges.empty()) {
                             // 截断当前区间到定义处
                             interval.ranges.front().start = id;
@@ -190,7 +190,7 @@ private:
                     if (use.isVirtual()) {
                         live.insert(use.vreg());
                         LiveInterval& interval = getInterval(use.vreg());
-                        interval.spillWeight += std::pow(10.0f, block->loopDepth());
+                        interval.spillWeight += static_cast<float>(std::pow(10.0, block->loopDepth()));
                         bool isSameAs = false;
                         for (size_t j = 0; j < inst->defs().size(); ++j) {
                             if (inst->defConstraints()[j].type == LIRConstraintType::SameAsInput && inst->defConstraints()[j].value == i) {
@@ -207,7 +207,7 @@ private:
                     if (use.isVirtual()) {
                         live.insert(use.vreg());
                         LiveInterval& interval = getInterval(use.vreg());
-                        interval.spillWeight += std::pow(10.0f, block->loopDepth());
+                        interval.spillWeight += static_cast<float>(std::pow(10.0, block->loopDepth()));
                         interval.addRange(blockStart, id + 1);
                     }
                 }
